4001 - Add GridHeadCell component
