+++ src/client/pages/Section/DataTable/Table/GridHead/GridHead.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { Col as TypeCol, Row as TypeRow, RowType, Table } from 'meta/assessment'
 import { RecordAssessmentData } from 'meta/data'
 
-import { useRenderCell } from './hooks/useRenderCell'
+import GridHeadCell from 'client/pages/Section/DataTable/Table/GridHead/GridHeadCell'
 
 type Props = {
   assessmentName: string
@@ -15,14 +15,24 @@ type Props = {
 const GridHead: React.FC<Props> = (props) => {
   const { assessmentName, data, headers, table } = props
 
-  const renderCell = useRenderCell({ assessmentName, data, headers, table })
   const rowsHeader = table.rows.filter((row) => row.props.type === RowType.header)
 
   return (
     <>
       {rowsHeader.map((row: TypeRow, rowIndex: number) =>
         row.cols.map((col: TypeCol, colIndex: number) => {
-          return renderCell({ col, colIndex, row, rowIndex })
+          return (
+            <GridHeadCell
+              assessmentName={assessmentName}
+              col={col}
+              colIndex={colIndex}
+              data={data}
+              headers={headers}
+              row={row}
+              rowIndex={rowIndex}
+              table={table}
+            />
+          )
         })
       )}
     </>
