+++ src/client/pages/Section/DataTable/Table/GridHead/GridHeadCell/index.ts
@@ -1,0 +1,1 @@
+export { default } from './GridHeadCell'
